Move getInvoiceItemsSchema to dedicated file
Refactored getInvoiceItemsSchema and its type from zuoraSchemas.ts to a new file, invoiceItem.ts, under types/objects for better modularity and organization.
